move core-tests into core/tests
