--- modules/core-tests/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@luma.gl/core-tests",
-  "version": "9.0.0-beta.7",
-  "description": "luma.gl internal test module",
-  "license": "MIT",
-  "type": "module",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/visgl/luma.gl"
-  },
-  "keywords": [
-    "webgl",
-    "visualization",
-    "animation",
-    "3d"
-  ],
-  "types": "dist/index.d.ts",
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "files": [
-    "src",
-    "dist",
-    "dist.min.js",
-    "README.md"
-  ],
-  "sideEffects": false,
-  "scripts": {
-    "pre-build": "echo test utils has no bundle"
-  },
-  "devDependencies": {
-    "@luma.gl/core": "9.0.0-beta.7",
-    "@luma.gl/engine": "9.0.0-beta.7",
-    "@luma.gl/shadertools": "9.0.0-beta.7",
-    "@luma.gl/test-utils": "9.0.0-beta.7",
-    "@luma.gl/webgl": "9.0.0-beta.7",
-    "@luma.gl/webgpu": "9.0.0-beta.7"
-  },
-  "gitHead": "c636c34b8f1581eed163e94543a8eb1f4382ba8e"
-}
