--- modules/core-tests/src/index.ts
@@ -1,5 +1,0 @@
-// luma.gl
-// SPDX-License-Identifier: MIT
-// Copyright (c) vis.gl contributors
-
-// intentionally empty, this is a test-only module
