--- modules/core-tests/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../tsconfig.module.json",
-  "include": ["src/**/*"],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "strict": true,
-    "noImplicitAny": true,
-    "composite": true,
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "references": [
-    {"path": "../core"},
-    {"path": "../test-utils"},
-    {"path": "../webgl"},
-    {"path": "../webgpu"}
-  ]
-}
