+++ modules/core/test/adapter/device-helpers/device-info.spec.ts
