+++ modules/core/test/adapter/device-helpers/set-device-parameters.spec.ts
