+++ modules/core/test/adapter/resources/buffer.spec.ts
