+++ modules/core/test/adapter/resources/compute-pipeline.spec.ts
