+++ modules/core/test/adapter/resources/query-set.spec.ts
