+++ modules/core/test/adapter/resources/render-pipeline.spec.ts
