+++ modules/core/test/adapter/resources/sampler.spec.ts
