+++ modules/core/test/adapter/resources/shader.spec.ts
