+++ modules/core/test/adapter/resources/texture.spec.ts
