+++ modules/core/test/adapter/resources/vertex-array.spec.ts
